+++ src/lib.rs
@@ -82,21 +82,21 @@ pub mod process;
 pub mod session;
 pub mod reader;
 
-pub use session::{spawn, spawn_bash, spawn_python};
+pub use session::{spawn, spawn_bash, spawn_python, spawn_stream};
+pub use reader::{Until};
 
 pub mod errors {
     use std::time;
-    use crate::process::wait;
     // Create the Error, ErrorKind, ResultExt, and Result types
     error_chain::error_chain!{
         errors {
-            EOF(expected:String, got:String, exit_code:Option<wait::WaitStatus>) {
+            EOF(expected:String, got:String, exit_code:Option<String>) {
                 description("End of filestream (usually stdout) occurred, most probably\
                              because the process terminated")
                 display("EOF (End of File): Expected {} but got EOF after reading \"{}\", \
                              process terminated with {:?}", expected, got,
-                             exit_code.map(|status| format!("{:?}", status))
-                             .unwrap_or("unknown".to_string()))
+                             exit_code.as_ref()
+                             .unwrap_or(& "unknown".to_string()))
             }
             BrokenPipe {
                 description("The pipe to the process is broken. Most probably because\
@@ -108,6 +108,10 @@ pub mod errors {
                 display("Timeout Error: Expected {} but got \"{}\" (after waiting {} ms)",
                         expected, got, (timeout.as_secs() * 1000) as u32
                         + timeout.subsec_nanos() / 1_000_000)
+            }
+            EmptyProgramName {
+                description("The provided program name is empty.")
+                display("EmptyProgramName")
             }
         }
     }